설정 페이지 대신 모달로 띄우도록 수정 (#143)
* remove(settings/index.tsx): 셋팅 페이지 삭제

* feat(RoomSettingScreen): roomSettingScreen 모달 스크린 컴포넌트 추가 & 적용
