+++ pages/rooms/[roomId]/index.tsx
@@ -15,6 +15,7 @@ import {
   Emoji,
 } from "~/components/domains";
 import QRCodeCard from "~/components/domains/QRCodeCard";
+import RoomSettingScreen from "~/components/domains/RoomSettingScreen";
 import {
   Modal,
   Spacer,
@@ -120,11 +121,16 @@ const RoomContentPage: NextPage<Props> = () => {
             leftIconButton={<TopBarIconButton iconName="star" boxSize={29} />}
             rightIconButton={
               roomData?.currentUser.role === "CREATOR" ? (
-                <TopBarIconButton
-                  iconName="setting"
-                  onClick={() => {
-                    router.push(`/rooms/${roomId}/setting`);
-                  }}
+                <Modal
+                  trigger={({ open }) => (
+                    <TopBarIconButton iconName="setting" onClick={open} />
+                  )}
+                  modal={({ close }) => (
+                    <RoomSettingScreen
+                      onClickBackButton={close}
+                      room={roomData}
+                    />
+                  )}
                 />
               ) : (
                 <></>
